Fixed rgbchannels, implemented using hsv
Turns out the reason using rgbchannels didnt really work that well,
appears to have been because it accidently used only the red channel.
Using hsv is now implemented. It appears to get more "details" than the
others, such as the bricks the wall is made of. It does however seem to
be bad for our use, since this would be noise to us, and it is owrse
than the rgbchannels one at text.