Implemented a paste of ISEF from the book
Cant upload because of licenses, as its not yet rewritten so noone can
contest it being theirs.
We wont use it anyway, as it did not work well for our tests, and is
unlikely to be possible to improve within our timeframe. Hence it wont
be rewritten and uploaded either.
Some nice imgutil stuff and such as well, which is what is actually
commited.